fix: 🐛 import paths and font variables based on refactoring
