+++ src/components/header/Header.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { translate } from '../../resources/scripts/i18n/translate';
+import { translate } from '../../utils/translate';
 import { Headline } from '../headline/Headline';
 import { Text } from '../text/Text';
 import './header.styles';
